Updated the member import
Empty locality doesn't get add to the database.
Moved member import into a new seeder & added call to the new seeder.
Created a migration to modify the database. personal_information foreign key for the locality can be null.
Fixed member display so that if a personal_information doesn't have a user they are still displayed.
Added "doctrine/dbal" to composer.json. This allows the update of columns in the database.